Fixed a file rename bug.
git-svn-id: svn://svn.code.sf.net/p/ifolder/code/trunk@4383 db2431b9-df06-0410-971c-a38ea9a75485

+++ simias/src/Sync/FileMonitor.cs
@@ -615,8 +615,12 @@ namespace Simias.Sync
 									sn = GetShallowNodeForFile(args.OldFullPath);
 									if (sn != null && sn.ID != node.ID)
 									{
-										node = collection.GetNodeByID(sn.ID);
-										DeleteNode(node);
+										// If the file no longer exists delet the node.
+										if (!File.Exists(args.OldFullPath))
+										{
+											node = collection.GetNodeByID(sn.ID);
+											DeleteNode(node);
+										}
 									}
 									break;
 								}
